+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added Crop operation.
 * Added Overlay operation.
 * Added Map operation.
+* Added Sequence operation.
 
 # 0.1.0 (2019-09-15)